100% coverage for integration tests
